Move `GrailsPlaceholderConfigurer` `GrailsBeanPropertyOverrideConfigurer` from grace-spring-boot to grace-plugin-core
