+++ grace-plugin-core/src/main/groovy/org/grails/spring/context/support/GrailsBeanPropertyOverrideConfigurer.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2009-2023 the original author or authors.
+ * Copyright 2009-2025 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
